Apply suggestions from code review.
