+++ src/script/lua_api/l_mainmenu.h
@@ -151,7 +151,6 @@ private:
 	// async
 	static int l_do_async_callback(lua_State *L);
 
-	
 public:
 
 	/**
